requirements.txt
flask
flask-sqlalchemy
flask-wtf
wtforms-alchemy
